util.cを複数ファイルに分割
